+++ src/mediathequeroubaix/auth/authenticate.py
@@ -1,27 +1,28 @@
+import requests
 from requests import Response, Session
 from returns.io import IOResultE, impure_safe
 from returns.pipeline import flow
 from returns.pointfree import bind_result
 from returns.result import safe
 
-from mediathequeroubaix.login.authenticated_session import AuthenticatedSession
-from mediathequeroubaix.login.get_user import get_user
+from mediathequeroubaix.auth.authenticated_session import AuthenticatedSession
+from mediathequeroubaix.auth.get_user import get_user
+from mediathequeroubaix.config import User
 
 
-def login(
-    session: Session, username: str, password: str
-) -> IOResultE[AuthenticatedSession]:
+def authenticate(user: User) -> IOResultE[AuthenticatedSession]:
+    session = requests.Session()
     return flow(
-        (session, username, password),
-        connect,
-        bind_result(get_html),
+        (requests.Session(), user.login, user.password),
+        _connect,
+        bind_result(_get_html),
         bind_result(get_user),
         bind_result(safe(lambda user: AuthenticatedSession(session, user))),
     )
 
 
 @impure_safe
-def connect(data: tuple[Session, str, str]) -> Response:
+def _connect(data: tuple[Session, str, str]) -> Response:
     session, username, password = data
     response = session.post(
         "http://www.mediathequederoubaix.fr/user",
@@ -36,5 +37,5 @@ def connect(data: tuple[Session, str, str]) -> Response:
 
 
 @safe
-def get_html(response: Response) -> str:
+def _get_html(response: Response) -> str:
     return response.text