+++ src/mediathequeroubaix/get_loans/get_loans_test.py
@@ -5,8 +5,8 @@ from requests_mock import mock
 from returns.pipeline import is_successful
 from returns.unsafe import unsafe_perform_io
 
+from mediathequeroubaix.auth.authenticated_session import AuthenticatedSession, Username
 from mediathequeroubaix.get_loans.get_loans import get_loans
-from mediathequeroubaix.login.authenticated_session import AuthenticatedSession, User
 
 
 def test_ok(requests_mock: mock) -> None:
@@ -21,12 +21,13 @@ def test_ok(requests_mock: mock) -> None:
     """,
     )
 
-    result = get_loans(AuthenticatedSession(requests.Session(), User("John doe")))
+    result = get_loans(AuthenticatedSession(requests.Session(), Username("John Doe")))
 
     assert is_successful(result)
     actual = unsafe_perform_io(result.unwrap())
-    assert len(actual) == 1
-    assert actual[0].title == "L'épreuve d'Hadès"
+    assert actual.username == "John Doe"
+    assert len(actual.items) == 1
+    assert actual.items[0].title == "L'épreuve d'Hadès"
 
 
 def test_no_token(requests_mock: mock) -> None:
@@ -34,6 +35,6 @@ def test_no_token(requests_mock: mock) -> None:
         "http://www.mediathequederoubaix.fr/espace_personnel", text="some html"
     )
 
-    result = get_loans(AuthenticatedSession(requests.Session(), User("John doe")))
+    result = get_loans(AuthenticatedSession(requests.Session(), Username("John doe")))
 
     assert is_successful(result) is False