+++ src/mediathequeroubaix/get_loans/loan.py
@@ -3,6 +3,8 @@ from datetime import date, datetime
 from pydantic import validator
 from pydantic.main import BaseModel
 
+from mediathequeroubaix.auth.authenticated_session import Username
+
 
 class Loan(BaseModel):
     # Vampi
@@ -17,9 +19,14 @@ class Loan(BaseModel):
     itemcallnumber: str
     # Boolean
     renewable: bool
-    # too_many
+    # Values: too_many, booked
     reasons_not_renewable: str | None
 
     @validator("date_due", pre=True)
     def parse_date_due(cls, value: str) -> date:
         return datetime.strptime(value, "%Y-%m-%dT%H:%M:%S").date()
+
+
+class Loans(BaseModel):
+    username: Username
+    items: list[Loan]